+++ uefi-raw/src/protocol/network/mod.rs
@@ -3,3 +3,4 @@ pub mod http;
 pub mod ip4;
 pub mod ip4_config2;
 pub mod tls;
+pub mod tcpv4;