+++ uefi-raw/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `DevicePathToTextProtocol` and `DevicePathFromTextProtocol`.
 - Added minor utility methods to `Ipv4Address`: `new(u8, u8, u8, u8) -> Self` and `zero() -> Self`.
 - Added `Ip4ModeData`, `Ip4ConfigData`, and `Ip4IcmpType`.
+- Added `TCPv4Option`, `TCPv4ConnectionState`, and `TCPv4ServiceBindingProtocol`.
 
 # uefi-raw - 0.5.1 (2024-03-17)